+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HCI: hci_remove_event_handler to remove packet handler
 - hci_cmd: support variable length fields and arrayed parameters
 - GAP: ENABLE_EXPLICIT_LINK_KEY_REPLY allows for asynchronous link key lookup by application
+- GAP: gap_set_page_timeout to set Page Timeout
 - POSIX: btstack_signal allows to register for callback on signal, e.g. ctrl-c
 - Windows: btstack_stdin_window_register_ctrl_c_callback allows to register for ctrl-c
 - A2DP: allow to register media codec validator for sink and source with: